+++ pkg/preference/implem.go
@@ -17,110 +17,8 @@ import (
 	"github.com/redhat-developer/odo/pkg/util"
 )
 
-const (
-	GlobalConfigEnvName  = "GLOBALODOCONFIG"
-	configFileName       = "preference.yaml"
-	preferenceKind       = "Preference"
-	preferenceAPIVersion = "odo.dev/v1alpha1"
-
-	//DefaultTimeout for openshift server connection check (in seconds)
-	DefaultTimeout = 1
-
-	// DefaultPushTimeout is the default timeout for pods (in seconds)
-	DefaultPushTimeout = 240
-
-	// DefaultBuildTimeout is the default build timeout for pods (in seconds)
-	DefaultBuildTimeout = 300
-
-	// UpdateNotificationSetting is the name of the setting controlling update notification
-	UpdateNotificationSetting = "UpdateNotification"
-
-	// UpdateNotificationSettingDescription is human-readable description for the update notification setting
-	UpdateNotificationSettingDescription = "Flag to control if an update notification is shown or not (Default: true)"
-
-	// NamePrefixSetting is the name of the setting controlling name prefix
-	NamePrefixSetting = "NamePrefix"
-
-	// NamePrefixSettingDescription is human-readable description for the name prefix setting
-	NamePrefixSettingDescription = "Use this value to set a default name prefix (Default: current directory name)"
-
-	// TimeoutSetting is the name of the setting controlling timeout for connection check
-	TimeoutSetting = "Timeout"
-
-	// BuildTimeoutSetting is the name of the setting controlling BuildTimeout
-	BuildTimeoutSetting = "BuildTimeout"
-
-	// PushTimeoutSetting is the name of the setting controlling PushTimeout
-	PushTimeoutSetting = "PushTimeout"
-
-	// RegistryCacheTimeSetting is human-readable description for the registrycachetime setting
-	RegistryCacheTimeSetting = "RegistryCacheTime"
-
-	// DefaultDevfileRegistryName is the name of default devfile registry
-	DefaultDevfileRegistryName = "DefaultDevfileRegistry"
-
-	// DefaultDevfileRegistryURL is the URL of default devfile registry
-	DefaultDevfileRegistryURL = "https://registry.devfile.io"
-
-	// OldDefaultDevfileRegistryURL is the URL of old default devfile registry for registry migration purpose
-	OldDefaultDevfileRegistryURL = "https://github.com/odo-devfiles/registry"
-
-	// DefaultRegistryCacheTime is time (in minutes) for how long odo will cache information from Devfile registry
-	DefaultRegistryCacheTime = 15
-
-	// EphemeralSetting specifies if ephemeral volumes needs to be used as source volume.
-	EphemeralSetting = "Ephemeral"
-
-	// DefaultEphemeralSettings is a default value for Ephemeral preference
-	DefaultEphemeralSettings = true
-
-	// ConsentTelemetrySettings specifies if the user consents to telemetry
-	ConsentTelemetrySetting = "ConsentTelemetry"
-
-	// DefaultConsentTelemetry is a default value for ConsentTelemetry preference
-	DefaultConsentTelemetrySetting = false
-)
-
-// TimeoutSettingDescription is human-readable description for the timeout setting
-var TimeoutSettingDescription = fmt.Sprintf("Timeout (in seconds) for OpenShift server connection check (Default: %d)", DefaultTimeout)
-
-// PushTimeoutSettingDescription adds a description for PushTimeout
-var PushTimeoutSettingDescription = fmt.Sprintf("PushTimeout (in seconds) for waiting for a Pod to come up (Default: %d)", DefaultPushTimeout)
-
-// BuildTimeoutSettingDescription adds a description for BuildTimeout
-var BuildTimeoutSettingDescription = fmt.Sprintf("BuildTimeout (in seconds) for waiting for a build of the git component to complete (Default: %d)", DefaultBuildTimeout)
-
-// RegistryCacheTimeDescription adds a description for RegistryCacheTime
-var RegistryCacheTimeDescription = fmt.Sprintf("For how long (in minutes) odo will cache information from Devfile registry (Default: %d)", DefaultRegistryCacheTime)
-
-// EphemeralDescription adds a description for EphemeralSourceVolume
-var EphemeralDescription = fmt.Sprintf("If true, odo will create an emptyDir volume to store source code (Default: %t)", DefaultEphemeralSettings)
-
-//TelemetryConsentDescription adds a description for TelemetryConsentSetting
-var ConsentTelemetryDescription = fmt.Sprintf("If true, odo will collect telemetry for the user's odo usage (Default: %t)\n\t\t    For more information: https://developers.redhat.com/article/tool-data-collection", DefaultConsentTelemetrySetting)
-
-// This value can be provided to set a seperate directory for users 'homedir' resolution
-// note for mocking purpose ONLY
-var customHomeDir = os.Getenv("CUSTOM_HOMEDIR")
-
-var (
-	// records information on supported parameters
-	supportedParameterDescriptions = map[string]string{
-		UpdateNotificationSetting: UpdateNotificationSettingDescription,
-		NamePrefixSetting:         NamePrefixSettingDescription,
-		TimeoutSetting:            TimeoutSettingDescription,
-		BuildTimeoutSetting:       BuildTimeoutSettingDescription,
-		PushTimeoutSetting:        PushTimeoutSettingDescription,
-		RegistryCacheTimeSetting:  RegistryCacheTimeDescription,
-		EphemeralSetting:          EphemeralDescription,
-		ConsentTelemetrySetting:   ConsentTelemetryDescription,
-	}
-
-	// set-like map to quickly check if a parameter is supported
-	lowerCaseParameters = util.GetLowerCaseParameters(GetSupportedParameters())
-)
-
 // odoSettings holds all odo specific configurations
+// these configurations are applicable across the odo components
 type odoSettings struct {
 	// Controls if an update notification is shown or not
 	UpdateNotification *bool `yaml:"UpdateNotification,omitempty"`
@@ -202,13 +100,13 @@ func newPreference() Preference {
 	}
 }
 
-// newPreferenceInfo gets the PreferenceInfo from preference file and creates the preference file in case it's
-// not present
+// newPreferenceInfo gets the PreferenceInfo from preference file
+// or returns default PreferenceInfo if preference file does not exist
 func newPreferenceInfo() (*preferenceInfo, error) {
 	preferenceFile, err := getPreferenceFile()
 	klog.V(4).Infof("The path for preference file is %+v", preferenceFile)
 	if err != nil {
-		return nil, errors.Errorf("unable to get odo preference file, run the command with more verbosity to check if the preference path is correct")
+		return nil, err
 	}
 
 	c := preferenceInfo{
@@ -259,7 +157,7 @@ func newPreferenceInfo() (*preferenceInfo, error) {
 func (c *preferenceInfo) RegistryHandler(operation string, registryName string, registryURL string, forceFlag bool, isSecure bool) error {
 	var registryList []Registry
 	var err error
-	registryExist := false
+	var registryExist bool
 
 	// Registry list is empty
 	if c.OdoSettings.RegistryList == nil {
@@ -351,8 +249,7 @@ func handleWithRegistryExist(index int, registryList []Registry, operation strin
 	return registryList, nil
 }
 
-// SetConfiguration modifies Odo configurations in the config file
-// as of now being used for nameprefix, timeout, updatenotification
+// SetConfiguration modifies odo preferences in the preference file
 // TODO: Use reflect to set parameters
 func (c *preferenceInfo) SetConfiguration(parameter string, value string) error {
 	if p, ok := asSupportedParameter(parameter); ok {
@@ -425,7 +322,7 @@ func (c *preferenceInfo) SetConfiguration(parameter string, value string) error 
 			c.OdoSettings.ConsentTelemetry = &val
 		}
 	} else {
-		return errors.Errorf("unknown parameter : %q is not a parameter in odo preference, run help to see list of available parameters", parameter)
+		return errors.Errorf("unknown parameter : %q is not a parameter in odo preference, run `odo preference -h` to see list of available parameters", parameter)
 	}
 
 	err := util.WriteToFile(&c.Preference, c.Filename)
@@ -435,8 +332,7 @@ func (c *preferenceInfo) SetConfiguration(parameter string, value string) error 
 	return nil
 }
 
-// DeleteConfiguration delete Odo configurations in the global config file
-// as of now being used for nameprefix, timeout, updatenotification
+// DeleteConfiguration deletes odo preference from the odo preference file
 func (c *preferenceInfo) DeleteConfiguration(parameter string) error {
 	if p, ok := asSupportedParameter(parameter); ok {
 		// processing values according to the parameter names
